backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0,<5
pydantic-settings==2.6.1
python-multipart==0.0.17
# Vertex AI RAG
google-cloud-aiplatform>=1.38.0
google-genai>=1.0.0
